First installment of merging upstreamm changes.
Update to gulp 4.0, and new versions of packages.
Add files needed by new gulpfile.js.
