+++ examples/github.html
@@ -1,0 +1,141 @@
+
+		<table>
+			<tr>
+				<td>
+					<a href="http://potree.org/potree/examples/animation_paths.html" target="_blank">
+						<img src="examples/thumbnails/animation_paths.png" width="100%" />
+					</a>
+				</td><td>
+					<a href="http://potree.org/potree/examples/annotations.html" target="_blank">
+						<img src="examples/thumbnails/annotations.png" width="100%" />
+					</a>
+				</td><td>
+					<a href="http://potree.org/potree/examples/annotation_hierarchy.html" target="_blank">
+						<img src="examples/thumbnails/annotation_hierarchy.png" width="100%" />
+					</a>
+				</td><td>
+					<a href="http://potree.org/potree/examples/ca13.html" target="_blank">
+						<img src="examples/thumbnails/ca13.png" width="100%" />
+					</a>
+				</td><td>
+					<a href="http://potree.org/potree/examples/cesium_ca13.html" target="_blank">
+						<img src="examples/thumbnails/cesium_ca13.png" width="100%" />
+					</a>
+				</td><td>
+					<a href="http://potree.org/potree/examples/cesium_retz.html" target="_blank">
+						<img src="examples/thumbnails/cesium_retz.png" width="100%" />
+					</a>
+				</td>
+			</tr>
+			<tr>
+				<th>Animation Path</th><th>Annotations</th><th>Hierarchical Annotations</th><th>CA13</th><th>Cesium CA13</th><th>Cesium Retz</th>
+			</tr><tr>
+				<td>
+					<a href="http://potree.org/potree/examples/cesium_sorvilier.html" target="_blank">
+						<img src="examples/thumbnails/cesium_sorvilier.png" width="100%" />
+					</a>
+				</td><td>
+					<a href="http://potree.org/potree/examples/clipping_volume.html" target="_blank">
+						<img src="examples/thumbnails/clipping_volume.png" width="100%" />
+					</a>
+				</td><td>
+					<a href="http://potree.org/potree/examples/custom_sidebar_section.html" target="_blank">
+						<img src="examples/thumbnails/custom_sidebar_section.png" width="100%" />
+					</a>
+				</td><td>
+					<a href="http://potree.org/potree/examples/elevation_profile.html" target="_blank">
+						<img src="examples/thumbnails/elevation_profile.png" width="100%" />
+					</a>
+				</td><td>
+					<a href="http://potree.org/potree/examples/embedded_iframe.html" target="_blank">
+						<img src="examples/thumbnails/embedded_iframe.png" width="100%" />
+					</a>
+				</td><td>
+					<a href="http://potree.org/potree/examples/gradient_colors.html" target="_blank">
+						<img src="examples/thumbnails/gradient_colors.png" width="100%" />
+					</a>
+				</td>
+			</tr>
+			<tr>
+				<th>Cesium Sorvilier</th><th>Clipping Volume</th><th>Custom Sidebar Section</th><th>Elevation Profile</th><th>Embedded Iframe</th><th>Gradient Colors</th>
+			</tr><tr>
+				<td>
+					<a href="http://potree.org/potree/examples/heidentor.html" target="_blank">
+						<img src="examples/thumbnails/heidentor.png" width="100%" />
+					</a>
+				</td><td>
+					<a href="http://potree.org/potree/examples/light_animated.html" target="_blank">
+						<img src="examples/thumbnails/light_animated.png" width="100%" />
+					</a>
+				</td><td>
+					<a href="http://potree.org/potree/examples/light_ca13.html" target="_blank">
+						<img src="examples/thumbnails/light_ca13.png" width="100%" />
+					</a>
+				</td><td>
+					<a href="http://potree.org/potree/examples/lines.html" target="_blank">
+						<img src="examples/thumbnails/lines.png" width="100%" />
+					</a>
+				</td><td>
+					<a href="http://potree.org/potree/examples/lion.html" target="_blank">
+						<img src="examples/thumbnails/lion.png" width="100%" />
+					</a>
+				</td><td>
+					<a href="http://potree.org/potree/examples/lion_las.html" target="_blank">
+						<img src="examples/thumbnails/lion_las.png" width="100%" />
+					</a>
+				</td>
+			</tr>
+			<tr>
+				<th>Heidentor</th><th>Animated Light</th><th>Light CA13</th><th>Lines</th><th>Lion</th><th>Lion LAS</th>
+			</tr><tr>
+				<td>
+					<a href="http://potree.org/potree/examples/lion_laz.html" target="_blank">
+						<img src="examples/thumbnails/lion_las.png" width="100%" />
+					</a>
+				</td><td>
+					<a href="http://potree.org/potree/examples/measurements.html" target="_blank">
+						<img src="examples/thumbnails/measurements.png" width="100%" />
+					</a>
+				</td><td>
+					<a href="http://potree.org/potree/examples/meshes.html" target="_blank">
+						<img src="examples/thumbnails/meshes.png" width="100%" />
+					</a>
+				</td><td>
+					<a href="http://potree.org/potree/examples/multiple_pointclouds.html" target="_blank">
+						<img src="examples/thumbnails/multiple_point_clouds.png" width="100%" />
+					</a>
+				</td><td>
+					<a href="http://potree.org/potree/examples/shapefiles.html" target="_blank">
+						<img src="examples/thumbnails/shapefiles.png" width="100%" />
+					</a>
+				</td><td>
+					<a href="http://potree.org/potree/examples/viewer.html" target="_blank">
+						<img src="examples/thumbnails/viewer.png" width="100%" />
+					</a>
+				</td>
+			</tr>
+			<tr>
+				<th>Lion LAZ</th><th>Measurements</th><th>Meshes</th><th>Multiple Point Clouds</th><th>Shapefiles</th><th>Basic Viewer</th>
+			</tr><tr>
+				<td>
+					<a href="http://potree.org/potree/examples/features_sorvilier.html" target="_blank">
+						<img src="examples/thumbnails/features_sorvilier.png" width="100%" />
+					</a>
+				</td><td>
+					<a href="http://potree.org/potree/examples/features_ca13.html" target="_blank">
+						<img src="examples/thumbnails/features_ca13.png" width="100%" />
+					</a>
+				</td><td>
+					<a href="http://potree.org/potree/examples/arena4d.html" target="_blank">
+						<img src="examples/thumbnails/arena4d.png" width="100%" />
+					</a>
+				</td><td>
+					<a href="http://potree.org/potree/examples/entwine_greyhound.html" target="_blank">
+						<img src="examples/thumbnails/entwine_greyhound.png" width="100%" />
+					</a>
+				</td>
+			</tr>
+			<tr>
+				<th>Features (Sorvilier)</th><th>Features (CA13)</th><th>Arena4D</th><th>Entwine / Greyhound</th>
+			</tr>
+		</table>
